requirements.txt
flask>=3.0.0
pytesseract>=0.3.10
Pillow>=10.4.0
deep-translator>=1.11.4
langdetect>=1.0.9
pdf2image>=1.17.0
reportlab>=4.2.0
werkzeug>=3.0.0
